requirements.txt
Pillow==11.2.1
pynput==1.8.1
pyperclip==1.9.0
pystray==0.19.5
tqdm==4.67.1
